SendViewModel: Use `PasteFromClipboardCommand` based on `EventTriggerBehavior` instead of explicit `GotFocus` event
