+++ WalletWasabi.Fluent/Views/Wallets/Send/SendView.axaml
@@ -148,7 +148,6 @@
           <Image Width="120" Source="avares://WalletWasabi.Fluent/Assets/TechnologyLogos/payjoin.png"
                  DockPanel.Dock="Right" IsVisible="{Binding IsPayJoin}" Margin="0 8 8 0" VerticalAlignment="Top" />
           <DualCurrencyEntryBox x:Name="amountTb"
-                                GotFocus="AmountTb_GotFocusAsync"
                                 KeyboardNavigation.IsTabStop="{Binding !IsFixedAmount}"
                                 IsReadOnly="{Binding IsFixedAmount}"
                                 AmountBtc="{Binding AmountBtc}"
@@ -166,6 +165,10 @@
                                            FalseValue="{x:Static PlacementMode.BottomEdgeAlignedLeft}" />
             </DualCurrencyEntryBox.Resources>
             <Interaction.Behaviors>
+              <EventTriggerBehavior EventName="GotFocus">
+                <InvokeCommandAction Command="{Binding PasteFromClipboardCommand}" CommandParameter="{Binding #amountTb}" />
+              </EventTriggerBehavior>
+
               <FlyoutSuggestionBehavior Target="{Binding #amountTb.LeftEntryBox}" Content="{Binding BitcoinContent}"
                                         EqualityComparer="{x:Static helpers:AmountStringComparer.Instance}"
                                         PlacementMode="{Binding ConversionReversed, Converter={StaticResource SuggestionPlacementConverter}}" />
